pyproject.toml
[project]
name = "ucloud2dida"
version = "0.1.0"
description = "同步北邮 UCloud 和课堂派待办到滴答清单"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dida365>=0.1.9",
    "python-dotenv>=1.0.0",
    "html2text>=2024.2.26",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ucloud2dida"]

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=24.0",
    "ruff>=0.1.0",
]
